requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
google-generativeai>=0.8.3
flask>=3.0.0
flask-cors>=4.0.0

